requirements.txt
ccxt
requests
flask
python-dotenv
schedule
